debug the solver pack
- CHG : change on the method to instantiated regions in SimpleONO class
- NEW : add new method in FDDomain to obtain trapping layer material
- FIX : fix a bug in TrapToGateTunneling solver, where the tunnel tags of all the interface vertices are set during solving tunneling problem of each slice

+++ SctmUtils.h
@@ -245,7 +245,7 @@ namespace SctmUtils
 		void WriteVector(vector<double> &vec1, vector<double> &vec2, const char *title = "title not assigned");
 		void WriteVector(vector<double> &vec1, vector<double> &vec2, vector<double> &vec3, const char *title = "title not assigned");
 		void WriteVector(vector<double> &vec1, vector<double> &vec2, vector<double> &vec3, vector<double> vec4, const char *title = "title not assigned");
-		void ReadVector(vector<double> &vec1, vector<double> &vec2, vector<double> &vec3);
+		void ReadVector(vector<int> &vec1, vector<double> &vec2, vector<double> &vec3);
 
 		void WriteLine(string &line);
